Move barcode_analysis.py -> barcodes/analysis.py, and update impacted code.
